renamed and moved relevance eval
